[BDD-3921] - refactoring according to PR review suggestions, writing javadocs
